Store lead info
L1 nodes don't get a NegInf key unless needed which can lead to cases
where we'll try to split a node that shouldn't be. This commit adds some
metadata saying if a node is lead by a "proper" lead (a key with height
greater than its level) and only splits a node when this is the case.

It's not a _great_ solution, but it's incremental progress.